# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Gila E-Commerce API
  description: API specifications for challenge e-commerce application including products CRUD, bulk CSV import, Redis shopping cart, and purchase checkout.
  version: 1.0.0
servers:
  - url: /
    description: Root URL prefix
security:
  - bearerAuth: []
paths:
...
  /api/v1/cart:
    get:
      tags:
        - Cart
      summary: Get the current customer's shopping cart (Customer only)
      operationId: getCart
      responses:
        '200':
          description: Shopping cart retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartDto'
        '401':
          description: Unauthorized JWT session
        '403':
          description: Forbidden (requires customer role)
    post:
      tags:
        - Cart
      summary: Add or update item quantity in the cart (Customer only)
      operationId: updateCartItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemRequestDto'
      responses:
        '200':
          description: Shopping cart updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartDto'
        '400':
          description: Invalid request payload or quantity limits exceeded
        '403':
          description: Forbidden (requires customer role)
    delete:
      tags:
        - Cart
      summary: Clear all items from the cart (Customer only)
      operationId: clearCart
      responses:
        '204':
          description: Cart cleared successfully
        '403':
          description: Forbidden (requires customer role)
  /api/v1/cart/items/{productId}:
    delete:
      tags:
        - Cart
      summary: Remove a specific item from the cart (Customer only)
      operationId: removeCartItem
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Item removed and updated cart returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartDto'
        '404':
          description: Product not found in cart
        '403':
          description: Forbidden (requires customer role)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductDto:
      type: object
      required:
        - name
        - price
        - stock
        - category
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
          format: int32
        category:
          type: string
...
    CartDto:
      type: object
      required:
        - items
        - totalPrice
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItemDto'
        totalPrice:
          type: number
          format: double
    CartItemDto:
      type: object
      required:
        - product
        - quantity
      properties:
        product:
          $ref: '#/components/schemas/ProductDto'
        quantity:
          type: integer
    CartItemRequestDto:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
...