# target/generated-sources/openapi/src/main/java/com/gila/ecommerce/api/CartApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.gila.ecommerce.api;

import com.gila.ecommerce.dto.CartDto;
import com.gila.ecommerce.dto.CartItemRequestDto;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-10T21:53:01.882557200-06:00[GMT-06:00]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Cart", description = "the Cart API")
public interface CartApi {

    /**
     * DELETE /api/v1/cart : Clear all items from the cart (Customer only)
     *
     * @return Cart cleared successfully (status code 204)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "clearCart",
        summary = "Clear all items from the cart (Customer only)",
        tags = { "Cart" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Cart cleared successfully"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/v1/cart"
    )
    
    ResponseEntity<Void> clearCart(
        
    );


    /**
     * GET /api/v1/cart : Get the current customer&#39;s shopping cart (Customer only)
     *
     * @return Shopping cart retrieved successfully (status code 200)
     *         or Unauthorized JWT session (status code 401)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "getCart",
        summary = "Get the current customer's shopping cart (Customer only)",
        tags = { "Cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Shopping cart retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartDto.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized JWT session"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/cart",
        produces = { "application/json" }
    )
    
    ResponseEntity<CartDto> getCart(
        
    );


    /**
     * DELETE /api/v1/cart/items/{productId} : Remove a specific item from the cart (Customer only)
     *
     * @param productId  (required)
     * @return Item removed and updated cart returned (status code 200)
     *         or Product not found in cart (status code 404)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "removeCartItem",
        summary = "Remove a specific item from the cart (Customer only)",
        tags = { "Cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Item removed and updated cart returned", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product not found in cart"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/v1/cart/items/{productId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<CartDto> removeCartItem(
        @Parameter(name = "productId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("productId") UUID productId
    );


    /**
     * POST /api/v1/cart : Add or update item quantity in the cart (Customer only)
     *
     * @param cartItemRequestDto  (required)
     * @return Shopping cart updated successfully (status code 200)
     *         or Invalid request payload or quantity limits exceeded (status code 400)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "updateCartItem",
        summary = "Add or update item quantity in the cart (Customer only)",
        tags = { "Cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Shopping cart updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payload or quantity limits exceeded"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/cart",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<CartDto> updateCartItem(
        @Parameter(name = "CartItemRequestDto", description = "", required = true) @Valid @RequestBody CartItemRequestDto cartItemRequestDto
    );

}